+++ gateway-api/src/main/java/com/gateway/service/impl/PaymentLoggingServiceImpl.java
@@ -1,5 +1,6 @@
 package com.gateway.service.impl;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import org.slf4j.Logger;
@@ -23,9 +24,10 @@ public class PaymentLoggingServiceImpl implements PaymentLoggingService {
 	@Async("paymentLoggingExecutor")
 	public void paymentLog(Date requestDatetime, Date responsetDatetime, String rrn, String stan, String responseCode,
 			String responseDescription, String cnic, String mobile, String name, String consumerNumber, String billerId,
-			double amount, double charges, String activity, String paymentRefNo, String billerNumber,
-			String transactionStatus, String address, double transactionFees, double taxAmount, double total, String channel,
-			String billStatus, String tranDate, String tranTime, String province,String tranAuthId) throws Exception {
+			BigDecimal amount, double charges, String activity, String paymentRefNo, String billerNumber,
+			String transactionStatus, String address, double transactionFees, double taxAmount, double total,
+			String channel, String billStatus, String tranDate, String tranTime, String province, String tranAuthId)
+			throws Exception {
 
 		PaymentLog paymentLog = new PaymentLog();
 		LOG.info("Inserting in table (paymentLog audit)");
